requirements.txt
customtkinter==5.2.2
GPUtil==1.4.0
opencv-python==4.10.0.84
pillow==11.0.0
pySmartDL==1.3.4
pystray==0.19.5
pyTelegramBotAPI==4.26.0
requests==2.31.0
uptime==3.0.1
